add api update account

+++ ship-convenient/Model/TransactionModel/ResponseTransactionModel.cs
@@ -5,8 +5,8 @@
         public string Status { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public string TransactionType { get; set; } = string.Empty;
-        public decimal CoinExchange { get; set; }
-        public decimal BalanceWallet { get; set; }
+        public int CoinExchange { get; set; }
+        public int BalanceWallet { get; set; }
         public DateTime CreatedAt { get; set; }
         public Guid? PackageId { get; set; } = null;
     }
